Server_Modify_ClientCheck
클라이언트 확인하는 패킷을 만들어서 농장을 들어갈때의 패킷에서 분리 시킴

+++ NeoSports_Server/Server_BOOST/ServerMain/GameMG.cpp
@@ -34,7 +34,7 @@ void GameMG::SetRopePos(float ropePos)
 	_ropePos += ropePos;
 }
 
-int GameMG::GetRopePos()
+float GameMG::GetRopePos()
 {
 	return _ropePos;
 }
