+++ NeoSports_Server/Server_BOOST/ServerMain/GameMG.h
@@ -20,5 +20,5 @@ public:
 	void Init();
 
 	void SetRopePos(float ropePos);
-	int GetRopePos();
+	float GetRopePos();
 };
